requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
environ==1.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
folium==0.12.1.post1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-dotenv==0.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
